Extract common baseclass for Bots
